--- modules/praxislive.graph/pom.xml
@@ -1,65 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <parent>
-    <groupId>org.praxislive.ide</groupId>
-    <artifactId>praxislive-parent</artifactId>
-    <version>6.0.0-SNAPSHOT</version>
-    <relativePath>../../pom.xml</relativePath>
-  </parent>
-
-  <artifactId>praxislive-graph</artifactId>
-  <packaging>nbm</packaging>
-  
-  <name>PraxisLIVE-Graph</name>
-
-  <properties>
-    <nbm.codeNameBase>org.praxislive.ide.graph</nbm.codeNameBase>
-  </properties>
-
-  <dependencies>
-    <dependency>
-      <groupId>org.netbeans.api</groupId>
-      <artifactId>org-netbeans-api-visual</artifactId>
-      <version>${netbeans.version}</version>
-      <type>jar</type>
-    </dependency>
-    <dependency>
-      <groupId>org.netbeans.api</groupId>
-      <artifactId>org-openide-util-lookup</artifactId>
-      <version>${netbeans.version}</version>
-      <type>jar</type>
-    </dependency>
-    <dependency>
-      <groupId>org.netbeans.api</groupId>
-      <artifactId>org-openide-util-ui</artifactId>
-      <version>${netbeans.version}</version>
-      <type>jar</type>
-    </dependency>
-  </dependencies>
-
-  <build>
-    <plugins>
-      <plugin>
-        <groupId>org.apache.netbeans.utilities</groupId>
-        <artifactId>nbm-maven-plugin</artifactId>
-        <configuration>
-          <codeNameBase>${nbm.codeNameBase}</codeNameBase>
-          <publicPackages>
-            <publicPackage>org.praxislive.ide.graph</publicPackage>
-          </publicPackages>
-        </configuration>
-        <extensions>true</extensions>
-      </plugin>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-jar-plugin</artifactId>
-        <configuration>
-          <archive>
-            <manifestFile>${project.build.outputDirectory}/META-INF/MANIFEST.MF</manifestFile>
-          </archive>
-        </configuration>
-      </plugin>
-    </plugins>
-  </build>
-</project>
