+++ modules/praxislive.pxr.graph/src/main/java/org/praxislive/ide/pxr/graph/GraphEditor.java
@@ -61,14 +61,14 @@ import org.praxislive.core.protocols.ContainerProtocol;
 import org.praxislive.ide.core.api.Callback;
 import org.praxislive.ide.core.api.Syncable;
 import org.praxislive.ide.core.ui.api.Actions;
-import org.praxislive.ide.graph.Alignment;
-import org.praxislive.ide.graph.EdgeID;
-import org.praxislive.ide.graph.EdgeWidget;
-import org.praxislive.ide.graph.NodeWidget;
-import org.praxislive.ide.graph.ObjectSceneAdaptor;
-import org.praxislive.ide.graph.PinID;
-import org.praxislive.ide.graph.PinWidget;
-import org.praxislive.ide.graph.PraxisGraphScene;
+import org.praxislive.ide.pxr.graph.scene.Alignment;
+import org.praxislive.ide.pxr.graph.scene.EdgeID;
+import org.praxislive.ide.pxr.graph.scene.EdgeWidget;
+import org.praxislive.ide.pxr.graph.scene.NodeWidget;
+import org.praxislive.ide.pxr.graph.scene.ObjectSceneAdaptor;
+import org.praxislive.ide.pxr.graph.scene.PinID;
+import org.praxislive.ide.pxr.graph.scene.PinWidget;
+import org.praxislive.ide.pxr.graph.scene.PraxisGraphScene;
 import org.praxislive.ide.model.ComponentProxy;
 import org.praxislive.ide.model.ContainerProxy;
 import org.praxislive.ide.model.RootProxy;
