+++ modules/praxislive.pxr.graph/src/main/java/org/praxislive/ide/pxr/graph/scene/PraxisConnectDecorator.java
@@ -19,7 +19,7 @@
  * Please visit https://www.praxislive.org if you need additional information or
  * have any questions.
  */
-package org.praxislive.ide.graph;
+package org.praxislive.ide.pxr.graph.scene;
 
 import java.awt.Point;
 import org.netbeans.api.visual.action.ConnectDecorator;
