+++ modules/praxislive.pxr.graph/src/main/resources/org/praxislive/ide/pxr/graph/scene/resources/vmd-collapse.png
